+++ Sources/Modules/Matrix/matrix.h
@@ -7,19 +7,19 @@
 #define MATRIX_H
 
 typedef struct {
-    uint8_t rows;
-    uint8_t cols;
+    uint16_t rows;
+    uint16_t cols;
     double * matrixData;
 } Matrix;
 
-Matrix * xCreateMatrix(uint8_t rows, uint8_t cols, double * newData);
+Matrix * xCreateMatrix(uint16_t rows, uint16_t cols, double * newData);
 Matrix * xDotProduct(Matrix * matrix1, Matrix * matrix2);
 Matrix * xMatrixAdd(Matrix * matrix1, Matrix * matrix2);
 Matrix * xMatrixSubtract(Matrix * matrix1, Matrix * matrix2);
 Matrix * xMatrixSquareEachElement(Matrix * m);
-static Matrix * xCreateEmptyMatrix(uint8_t rows, uint8_t cols);
-static double * xGetRow(Matrix * m, uint8_t rowNum);
-static double * xGetCol(Matrix * m, uint8_t colNum);
+static Matrix * xCreateEmptyMatrix(uint16_t rows, uint16_t cols);
+static double * xGetRow(Matrix * m, uint16_t rowNum);
+static double * xGetCol(Matrix * m, uint16_t colNum);
 
 void vPrintMatrix(Matrix * m);
